+++ hm.js
@@ -8,6 +8,7 @@ import { HMSpellItem } from './modules/item/spell-item.js';
 import { HMItemFactory } from './modules/item/item-factory.js';
 import { HMItemSheet } from './modules/item/item-sheet.js';
 import { HMArmorItemSheet } from './modules/item/armor-item-sheet.js';
+import { HMClassItemSheet } from './modules/item/class-item-sheet.js';
 import { HMRaceItemSheet } from './modules/item/race-item-sheet.js';
 import { HMWeaponItemSheet } from './modules/item/weapon-item-sheet.js';
 import { HMChatMgr } from './modules/mgr/chatmgr.js';
@@ -52,6 +53,7 @@ Hooks.once('init', async () => {
     Items.unregisterSheet('core', ItemSheet);
     Items.registerSheet('hackmaster', HMItemSheet, {makeDefault: true});
     Items.registerSheet('hackmaster', HMArmorItemSheet, {types: ['armor'], makeDefault: true});
+    Items.registerSheet('hackmaster', HMClassItemSheet, {types: ['cclass'], makeDefault: true});
     Items.registerSheet('hackmaster', HMRaceItemSheet, {types: ['race'], makeDefault: true});
     Items.registerSheet('hackmaster', HMWeaponItemSheet, {types: ['weapon'], makeDefault: true});